Move main_window to `src/gui`.
